wandra-backend/requirements.txt
django==3.2
djangorestframework
pyyaml
requests
django-cors-headers
